Add better valid staking lock check
Allows a user to deposit for a second time, as long as the duration of the total staking period is the same or greater.

This means that a user can stake for a shorter amount of time, as long as the ending lock time is greater or equal that the existing one.